pyproject.toml
[tool.poetry]
name = "innotter"
version = "0.1.0"
description = ""
authors = ["marinaRomanchuk"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
python-dotenv = "^0.20.0"
django-cors-headers = "^3.12.0"
PyJWT = "^2.4.0"
environs = "^9.5.0"
pre-commit = "^2.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
